backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
holidays>=0.35
python-dotenv>=1.0.0
psutil>=5.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiosmtplib>=3.0.0
sentry-sdk>=1.40.0
